pyproject.toml
[tool.poetry]
name = "we-worker"
version = "0.1.0"
description = "Worker component for Workflow Executor / Graph Data Explorer aimed at running on worker nodes to run tasks"
authors = ["Kashif Iftikhar <kashif@compulife.com.pk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
typer = "^0.9.0"
rich = "^13.7.1"
celery = "^5.3.6"
pika = "^1.3.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
ruff = "^0.3.3"
ipython = "^8.22.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
we-worker = 'we_worker.cli:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
src = ["we_worker"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--cov=we_worker --cov-fail-under=50"
testpaths = [
    "tests",
]

filterwarnings = [
    "ignore::DeprecationWarning",
]
